# apps/auth-service/src/main/java/com/cadentia/auth/api/RegisterRequest.java
package com.cadentia.auth.api;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @Email @NotBlank String email,
        @NotBlank @Size(min = 2, max = 120) String displayName,
        @NotBlank @Size(min = 12, max = 200) String password) {
}