# apps/auth-service/src/main/java/com/cadentia/auth/api/ResetPasswordRequest.java
package com.cadentia.auth.api;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record ResetPasswordRequest(
        @NotBlank String token,
        @NotBlank @Size(min = 12, max = 200) String newPassword) {
}