# apps/auth-service/src/main/java/com/cadentia/auth/api/TokenResponse.java
package com.cadentia.auth.api;

public record TokenResponse(
        String accessToken,
        String tokenType,
        long expiresIn,
        String refreshToken,
        AuthUserResponse user) {
}