# apps/auth-service/src/main/openapi/auth-service.yaml
openapi: 3.0.3
info:
  title: Cadentia First-Party Authentication Service
  version: 0.1.0
  description: Password authentication, JWT access tokens, refresh sessions, and self-service credential recovery.
servers:
  - url: http://localhost:8081
paths:
  /auth/login:
    post:
      operationId: login
      summary: Authenticate with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/register:
    post:
      operationId: register
      summary: Register a user when public registration is enabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '403':
          $ref: '#/components/responses/Forbidden'
  /auth/refresh:
    post:
      operationId: refresh
      summary: Rotate a refresh session and issue a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Rotated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /auth/logout:
    post:
      operationId: logout
      summary: Revoke a refresh session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Session revoked
  /auth/me:
    get:
      operationId: currentUser
      summary: Read the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
  /auth/password/change:
    post:
      operationId: changePassword
      summary: Change the current password and revoke existing sessions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed
  /auth/password/forgot:
    post:
      operationId: forgotPassword
      summary: Request a password reset without revealing account existence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Request accepted
  /auth/password/reset:
    post:
      operationId: resetPassword
      summary: Complete a password reset or account activation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password reset
  /internal/users/invitations:
    post:
      operationId: createInternalUserInvitation
      summary: Create a first-party account and one-time activation token for a trusted backend service
      security:
        - internalApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalInvitationRequest'
      responses:
        '201':
          description: Account and one-time activation token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalAuthInvitation'
        '400':
          description: Invitation data is invalid
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: An account already exists for the email
  /internal/users/by-email:
    get:
      operationId: findInternalUserByEmail
      summary: Resolve an active auth account for a trusted backend service
      security:
        - internalApiKey: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Active account identity without credential data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No active account exists for the email
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: X-Cadentia-Internal-Key
  responses:
    Unauthorized:
      description: Authentication failed
    Forbidden:
      description: Operation is disabled by configuration
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    RegisterRequest:
      type: object
      required: [email, displayName, password]
      properties:
        email: { type: string, format: email }
        displayName: { type: string, minLength: 2, maxLength: 120 }
        password: { type: string, format: password, minLength: 12 }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, format: password }
        newPassword: { type: string, format: password, minLength: 12 }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token: { type: string }
        newPassword: { type: string, format: password, minLength: 12 }
    InternalInvitationRequest:
      type: object
      required: [email, displayName]
      properties:
        email: { type: string, format: email }
        displayName: { type: string, minLength: 2, maxLength: 120 }
    InternalAuthInvitation:
      type: object
      required: [userId, email, displayName, activationToken]
      properties:
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        activationToken: { type: string, description: One-time token for the public password reset endpoint. }
    AuthUser:
      type: object
      required: [userId, email, displayName]
      properties:
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
    TokenResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, refreshToken, user]
      properties:
        accessToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer, format: int64 }
        refreshToken: { type: string }
        user:
          $ref: '#/components/schemas/AuthUser'